[MOB-35468] Выносит конфигурацию шаблонов в отдельные файлы для последующего переиспользования
