+++ Sources/FigmaGen/Models/Configuration/Templates/TemplateWrapper.swift
@@ -1,0 +1,22 @@
+import Foundation
+
+struct TemplateWrapper: Decodable {
+
+    // MARK: - Instance Properties
+
+    let templates: [Template]?
+
+    // MARK: - Initializers
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+
+        if container.decodeNil() {
+            self.templates = nil
+        } else if let singleValue = try? container.decode(Template.self) {
+            self.templates = [singleValue]
+        } else {
+            self.templates = try container.decode([Template].self)
+        }
+    }
+}
